Add text stroke, increase shadow size limit
This commit adds option to have sharp text border to make memes look more
canonical, and also increases shadow size limit to 100 because it is
useful for big images and text sizes. Shadow width was also renamed to
shadow size in the ui because i think it makes more sense, and drawCanvas
function is simplified a bit.